refactor(project-tracker): colocate tracker-specific modules
